refactor(stream-client): Implement review comments of PR#1139
