+++ apps/client/src/i18n/en.json
@@ -74,12 +74,10 @@
   "CANVAS:DELETE": "Delete",
   "CANVAS:FLOW-PARAMS-INPUT-OUTPUT": "Input Output",
   "CANVAS:FLOW-NAME-EXISTS":"Flow with name \"{{value}}\" already exists, please choose another name.",
-  "CANVAS:STREAM-NAME-EXISTS":"Stream with name \"{{value}}\" already exists, please choose another name.",
   "CANVAS:ERROR-GETTING-FLOW-NAME": "Error getting flow name",
   "CANVAS:ERROR-MESSAGE-FAILED": "Flow has failed with error code 700",
   "CANVAS:ERROR-MESSAGE": "Oops! something went wrong! :(",
   "CANVAS:ERROR-MESSAGE-UPDATE": "Update flow's {{value}} error",
-  "CANVAS:ERROR-MESSAGE-UPDATE-STREAM": "Update stream's {{value}} error",
   "CANVAS:FLOW-CANCELED": "Flow has been cancelled.",
   "CANVAS:RUN-ERROR:MAX-REACHED": "Timeout: flow didn't complete running.",
   "CANVAS:RUN-ERROR:RUN-FAILED": "An error occurred during flow execution.",
@@ -87,7 +85,6 @@
   "CANVAS:SHOW-INSTRUCTIONS": "Show instructions",
   "CANVAS:SUCCESS-MESSAGE-COMPLETED": "Flow completed! ^_^",
   "CANVAS:SUCCESS-MESSAGE-UPDATE": "Flow's {{value}} successfully updated",
-  "CANVAS:SUCCESS-MESSAGE-UPDATE-STREAM": "Stream's {{value}} successfully updated",
   "CANVAS:WARNING-UNSUPPORTED-TEST-RUN": "Running flows with iterator or subflow activities is not supported yet.",
   "TRIGGER-MENU:DELETE": "Delete",
   "CATEGORY-SELECTOR:CATEGORY": "CATEGORY",
@@ -286,5 +283,10 @@
   "BRANCH-CONFIGURATOR:CANCEL": "Cancel",
   "BRANCH-CONFIGURATOR:AVAILABLE-DATA": "Available Data",
   "PLUGIN-FLOW:TRIGGERS:DELETE-CONFIRMATION-TITLE": "Delete trigger",
-  "PLUGIN-FLOW:TRIGGERS:DELETE-CONFIRMATION-MESSAGE": "Are you sure you want to delete this trigger?"
+  "PLUGIN-FLOW:TRIGGERS:DELETE-CONFIRMATION-MESSAGE": "Are you sure you want to delete this trigger?",
+  "STREAMS": {
+    "DESIGNER:ERROR-STREAM-NAME-EXISTS": "Stream with name \"{{value}}\" already exists, please choose another name.",
+    "DESIGNER:ERROR-UPDATE-STREAM": "Update stream's {{value}} error",
+    "DESIGNER:SUCCESS-UPDATE-STREAM": "Stream's {{value}} successfully updated"
+  }
 }
