+++ libs/plugins/stream-client/src/lib/stream-client.module.ts
@@ -5,6 +5,7 @@ import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
 import { HeaderModule as FlogoDesignerHeader } from '@flogo-web/lib-client/designer-header';
 import { LogsModule as FlogoLogsModule } from '@flogo-web/lib-client/logs';
+import { SharedModule as FlogoSharedModule } from '@flogo-web/lib-client/common';
 import { StreamService, StreamSaveEffects } from './core';
 import { StreamDesignerComponent } from './stream-designer';
 import { StreamDataResolver } from './stream-data.resolver';
@@ -15,6 +16,7 @@ import { streamReducer } from './core/state';
     CommonModule,
     FlogoDesignerHeader,
     FlogoLogsModule,
+    FlogoSharedModule,
     StoreModule.forFeature('stream', streamReducer),
     EffectsModule.forFeature([StreamSaveEffects]),
     RouterModule.forChild([
