+++ libs/plugins/stream-client/src/lib/stream-designer/stream-designer.component.html
@@ -32,8 +32,7 @@
       >
         <ul class="dropdown-list">
           <li class="dropdown-list__item" (click)="deleteStream()">
-            <!--<i class="flogo-icon flogo-icon-delete"></i>{{ 'CANVAS:DELETE' | translate }}-->
-            <i class="flogo-icon flogo-icon-delete"></i>Delete
+            <i class="flogo-icon flogo-icon-delete"></i>{{ 'CANVAS:DELETE' | translate }}
           </li>
         </ul>
       </div>
